Update rss-to-bluesky.py
In this code, the trim_text() function is added to trim the text content of the post to the length of 300 graphemes before posting. The function is then used inside post_text() to trim the text content of the post before posting. You will need the uniseg library to do this, which you can install using pip: pip install uniseg.